App/Backend-Django/pyproject.toml
[tool.poetry]
name = "backend-django"
version = "0.1.0"
description = ""
authors = ["kkf017 <cpunk1017@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
markdown = "^3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
